+++ src/services/ServiceManager.ts
@@ -5,10 +5,15 @@ import {IDisposable} from '../utils';
 import StatusbarService from './StatusbarService';
 import GutterService from './GutterService';
 import WorkspaceService from './WorkspaceService';
+import GitService from './GitService';
 
 export interface IServiceManager {
   readonly configService: ConfigService;
   readonly decorationService: DecorationService;
+  readonly gutterService: GutterService;
+  readonly workspaceService: WorkspaceService;
+  readonly gitService: GitService;
+  readonly statusbarService: StatusbarService | undefined;
 }
 
 let _serviceManager: ServiceManager;
@@ -19,7 +24,7 @@ export class ServiceManager implements IServiceManager, IDisposable {
 
   readonly gutterService: GutterService;
   readonly workspaceService: WorkspaceService;
-
+  readonly gitService: GitService;
   private _statusbarService: StatusbarService | undefined;
   public get statusbarService(): StatusbarService | undefined {
     return this._statusbarService;
@@ -36,6 +41,7 @@ export class ServiceManager implements IServiceManager, IDisposable {
     this.gutterService = new GutterService(this);
     this.decorationService = new DecorationService(this);
     this.workspaceService = new WorkspaceService(this);
+    this.gitService = new GitService(this);
   }
 
   registerStatusbarService() {
@@ -45,15 +51,32 @@ export class ServiceManager implements IServiceManager, IDisposable {
   dispose(): void {
     this.configService.dispose();
     this.decorationService.dispose();
+    this.workspaceService.dispose();
+    this.gutterService.dispose();
+    this.gitService.dispose();
   }
 }
 
-export function initServiceManager(context: ExtensionContext) {
-  _serviceManager = new ServiceManager(context);
+export function initServiceManager(
+  context: ExtensionContext,
+  udpateCb: () => void,
+): Promise<ServiceManager> {
+  return new Promise((resolve, reject) => {
+    try {
+      _serviceManager = new ServiceManager(context);
+      _serviceManager.configService.onDecorationConfigChange(() => {
+        udpateCb();
+      });
+      resolve(_serviceManager);
+    } catch (error) {
+      reject(error);
+    }
+  });
 }
 
-export function postInitController(context: ExtensionContext) {
+export function postInitController() {
   _serviceManager.decorationService.setupAllDecorations();
+  _serviceManager.registerStatusbarService();
 }
 
 const resolveServiceManager = () => _serviceManager;